fix gbsinfo crash with many subsongs
See bug report at http://www.forallsecure.com/bug-reports/e917758134cc7d0f48ca0af77fcad861f1d36aa9/
252 subsongs got parsed as a negative number which made aus malloc() a negative amount of memory...

fix:
change data types of gbs.version, gbs.songs and gbs.dafaultsong from long to uint8_t

+++ gbs.h
@@ -1,7 +1,8 @@
 /*
  * gbsplay is a Gameboy sound player
  *
- * 2003-2005 (C) by Tobias Diedrich <ranma+gbsplay@tdiedrich.de>
+ * 2003-2005,2013 (C) by Tobias Diedrich <ranma+gbsplay@tdiedrich.de>
+ *                       Christian Garbs <mitch@cgarbs.de>
  * Licensed under GNU GPL.
  */
 
@@ -25,9 +26,9 @@ struct gbs_subsong_info {
 
 struct gbs {
 	char *buf;
-	long version;
-	long songs;
-	long defaultsong;
+	uint8_t version;
+	uint8_t songs;
+	uint8_t defaultsong;
 	uint16_t load;
 	uint16_t init;
 	uint16_t play;
@@ -53,7 +54,7 @@ struct gbs {
 	int16_t lmin, lmax, lvol, rmin, rmax, rvol;
 	long subsong_timeout, silence_timeout, fadeout, gap;
 	long long silence_start;
-	long subsong;
+	int subsong;
 	gbs_nextsubsong_cb nextsubsong_cb;
 	void *nextsubsong_cb_priv;
 };
